pyproject.toml
[tool.poetry]
name = "yggdrasilframework"
version = "0.1.0"
description = ""
authors = ["Silverteal <67404722+Silverteal@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "~0.112.2"
uvicorn = "~0.30.6"
pycryptodome = "~3.20.0"
pydantic = "~2.8.2"
starlette = "~0.38.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
